Day 3: Guard against edge case in input
I was lazy and messing around with math and forgot that you can
match every single line in the output. This would have made my integer
division of frequency // (len/2) == 2 instead of 1. Added in some
min's to fix this. Probably would replace this with an if statement if
I was working on this for real and not for fun.